Create movement code
The character speed continually increases when I try and add in the keys = pygame.key.get_pressed function, which I'm trying to add to make the movement smoother. The bug is shown when you hold either 'a' or 'd' to move left or right. He will continually increase his speed. I have also added the tests I have tried to fix this issue.