Add files via upload
All the assets you should need for the game

Also, you move left and right with 'a' and 'd' and jump with 'w'.
Your goal is to collect all the parts for the car and escape. You win when the car disappears. Congrats, you have beaten the game.
Some important things to remember, you are part mountain goat and your head is sticky. If this doesn't make at first, I hope it will later.